+++ src/components/text-label/text-label.module.css.d.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2011-2019 Genestack Limited
+ * All Rights Reserved
+ * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF GENESTACK LIMITED
+ * The copyright notice above does not evidence any
+ * actual or intended publication of such source code.
+ */
+export const root: string;
+export const inline: string;
+export const label: string;
+export const wrap: string;
+export const caption: string;